Corrections per review comments.
